Move device repo to repo package
